# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bibliotheque Num API
  version: 2026-02-16
  description: API backend complète (état actuel du code)
servers:
  - url: http://localhost:3001
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      responses:
        '200': { description: OK }
    patch:
      tags: [Users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name: { type: string }
                language: { type: string, enum: [fr, en, es, pt] }
                avatar_url: { type: string }
      responses:
        '200': { description: OK }
  /users/me/password:
    put:
      tags: [Users]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 8 }
      responses:
        '200': { description: OK }
        '401': { description: Current password invalid }
        '422': { description: Validation error (too short / same as current) }
  /users/me/avatar:
    post:
      tags: [Users]
      summary: Upload current user avatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        '200': { description: Avatar uploaded and profile updated }
        '400': { description: Missing file }
        '401': { description: Unauthorized }
        '422': { description: Invalid file type/size }
        '503': { description: Storage not configured }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT